+++ nengo/_vendor/npconv2d/conv2d.py
@@ -75,14 +75,9 @@ def extract_sliding_windows(x, ksize, pad, stride, floor_first=True):
     Returns:
         y: [N, (H-KH+PH+1)/SH, (W-KW+PW+1)/SW, KH * KW, C]
     """
-    n = x.shape[0]
-    h = x.shape[1]
-    w = x.shape[2]
-    c = x.shape[3]
-    kh = ksize[0]
-    kw = ksize[1]
-    sh = stride[0]
-    sw = stride[1]
+    n, h, w, c = x.shape
+    kh, kw = ksize
+    sh, sw = stride
 
     h2 = int(calc_size(h, kh, pad, sh))
     w2 = int(calc_size(w, kw, pad, sw))
@@ -112,6 +107,67 @@ def extract_sliding_windows(x, ksize, pad, stride, floor_first=True):
     return y
 
 
+def extract_sliding_windows_gradx(x,
+                                  ksize,
+                                  pad,
+                                  stride,
+                                  orig_size,
+                                  floor_first=False):
+    """Extracts windows on a dilated image.
+
+    Args:
+        x: [N, H', W', C] (usually dy)
+        k: [KH, KW]
+        pad: [PH, PW]
+        stride: [SH, SW]
+        orig_size: [H, W]
+
+    Returns:
+        y: [N, H, W, KH, KW, C]
+    """
+    n, h, w, c = x.shape
+    kh, kw = ksize
+    sh, sw = stride
+    ph, pw = pad
+    sh, sw = stride
+    h2, w2 = orig_size
+
+    xs = np.zeros([n, h, sh, w, sw, c])
+    xs[:, :, 0, :, 0, :] = x
+    xss = xs.shape
+    x = xs.reshape([xss[0], xss[1] * xss[2], xss[3] * xss[4], xss[5]])
+    x = x[:, :h2, :w2, :]
+
+    ph2 = int(np.ceil(ph / 2))
+    ph3 = int(np.floor(ph / 2))
+    pw2 = int(np.ceil(pw / 2))
+    pw3 = int(np.floor(pw / 2))
+    if floor_first:
+        pph = (ph3, ph2)
+        ppw = (pw3, pw2)
+    else:
+        pph = (ph2, ph3)
+        ppw = (pw2, pw3)
+    x = np.pad(
+        x, ((0, 0), pph, ppw, (0, 0)),
+        mode='constant',
+        constant_values=(0.0, ))
+
+    # The following code extracts window without copying the data:
+    # y = np.zeros([n, h2, w2, kh, kw, c])
+    # for ii in range(h2):
+    #     for jj in range(w2):
+    #         y[:, ii, jj, :, :, :] = x[:, ii:ii + kh, jj:jj + kw, :]
+    x_sn, x_sh, x_sw, x_sc = x.strides
+    y_strides = (x_sn, x_sh, x_sw, x_sh, x_sw, x_sc)
+    y = np.ndarray((n, h2, w2, kh, kw, c),
+                   dtype=x.dtype,
+                   buffer=x.data,
+                   offset=array_offset(x),
+                   strides=y_strides)
+    return y
+
+
 def conv2d(x, w, pad='SAME', stride=(1, 1)):
     """2D convolution (technically speaking, correlation).
     Args:
@@ -131,3 +187,48 @@ def conv2d(x, w, pad='SAME', stride=(1, 1)):
     y = x.dot(w)
     y = y.reshape([xs[0], xs[1], xs[2], -1])
     return y
+
+
+def conv2d_gradx(w, dy, xsize, pad='SAME', stride=(1, 1)):
+    """2D convolution gradient wrt. input.
+
+    Args:
+        dy: [N, H', W', K]
+        w: [I, J, K, C]
+        xsize: Original image size, [H, W]
+
+    Returns:
+        dx: [N, H, W, C]
+
+    Modifications from original code:
+      - Do not transpose input/output channels in `w`
+      - Fix bug in computing `pad2w` (use `dys[1]` instead of `dys[0]`)
+    """
+    assert w.shape[-2] == dy.shape[-1]
+    ksize = w.shape[:2]
+
+    if pad == 'SAME':
+        dys = dy.shape[1:3]
+        pad2h = int(
+            calc_pad('SAME', max(dys[0], dys[0] * stride[0] - 1), xsize[0], 1,
+                     ksize[0]))
+        pad2w = int(
+            calc_pad('SAME', max(dys[1], dys[1] * stride[1] - 1), xsize[1], 1,
+                     ksize[1]))
+        pad2 = (pad2h, pad2w)
+    elif pad == 'VALID':
+        pad2 = (int(calc_pad('SAME', 0, 0, 1, ksize[0])),
+                int(calc_pad('SAME', 0, 0, 1, ksize[1])))
+        pad2 = (pad2[0] * 2, pad2[1] * 2)
+    else:
+        pad2 = pad
+
+    ksize = w.shape[:2]
+    dx = extract_sliding_windows_gradx(dy, ksize, pad2, stride, xsize)
+    dxs = dx.shape
+    dx = dx.reshape([dxs[0] * dxs[1] * dxs[2], -1])
+    w = w[::-1, ::-1, :, :]
+    ws = w.shape
+    w = w.reshape([ws[0] * ws[1] * ws[2], ws[3]])
+    dx = dx.dot(w)
+    return dx.reshape([dxs[0], dxs[1], dxs[2], -1])