+++ nengo/builder/transforms.py
@@ -6,7 +6,13 @@ from nengo.builder.operator import DotInc, ElementwiseInc, Operator, Reset, Spar
 from nengo.builder.signal import Signal
 from nengo.exceptions import BuildError
 from nengo.rc import rc
-from nengo.transforms import Convolution, Dense, NoTransform, Sparse
+from nengo.transforms import (
+    Convolution,
+    ConvolutionTranspose,
+    Dense,
+    NoTransform,
+    Sparse,
+)
 
 
 def multiply(x, y):
@@ -79,6 +85,7 @@ def build_sparse(model, transform, sig_in, decoders=None, encoders=None, rng=np.
 
 
 @Builder.register(Convolution)
+@Builder.register(ConvolutionTranspose)
 def build_convolution(
     model, transform, sig_in, decoders=None, encoders=None, rng=np.random
 ):
@@ -100,8 +107,11 @@ def build_convolution(
     weighted = Signal(shape=transform.size_out, name="%s.weighted" % transform)
     model.add_op(Reset(weighted))
 
+    op_class = (
+        ConvTransposeInc if isinstance(transform, ConvolutionTranspose) else ConvInc
+    )
     model.add_op(
-        ConvInc(
+        op_class(
             weight_sig, sig_in, weighted, transform, tag="%s.apply_weights" % transform
         )
     )
@@ -109,10 +119,10 @@ def build_convolution(
     return weighted, weight_sig
 
 
-class ConvInc(Operator):
+class GeneralConvInc(Operator):
     """Apply convolutional weights to input signal.
 
-    .. versionadded:: 3.0.0
+    .. versionadded:: 3.2.0
 
     Parameters
     ----------
@@ -122,8 +132,8 @@ class ConvInc(Operator):
         The input signal.
     Y : Signal
         Output signal to be incremented.
-    conv : `~nengo.Convolution`
-        The Convolution object being applied.
+    conv : `~nengo.Convolution` or `~nengo.transforms.ConvolutionTranspose`
+        The Convolution or ConvolutionTranspose transform being applied.
     tag : str, optional
         A label associated with the operator, for debugging purposes.
 
@@ -135,8 +145,8 @@ class ConvInc(Operator):
         The input signal.
     Y : Signal
         Output signal to be incremented.
-    conv : `~nengo.Convolution`
-        The Convolution object being applied.
+    conv : `~nengo.Convolution` or `~nengo.transforms.ConvolutionTranspose`
+        The Convolution or ConvolutionTranspose transform being applied.
     tag : str, optional
         A label associated with the operator, for debugging purposes.
 
@@ -149,6 +159,8 @@ class ConvInc(Operator):
     """
 
     def __init__(self, W, X, Y, conv, tag=None):
+        assert isinstance(conv, (Convolution, ConvolutionTranspose))
+
         super().__init__(tag=tag)
 
         self.conv = conv
@@ -157,6 +169,10 @@ class ConvInc(Operator):
         self.incs = [Y]
         self.reads = [W, X]
         self.updates = []
+
+    @property
+    def transpose(self):
+        return isinstance(self.conv, ConvolutionTranspose)
 
     @property
     def W(self):
@@ -172,7 +188,8 @@ class ConvInc(Operator):
 
     @property
     def _descstr(self):
-        return "conv2d(%s, %s) -> %s" % (self.W, self.X, self.Y)
+        name = "convtranspose2d" if self.transpose else "conv2d"
+        return "%s(%s, %s) -> %s" % (name, self.W, self.X, self.Y)
 
     def make_step(self, signals, dt, rng):
         if self.conv.dimensions > 2:
@@ -185,6 +202,9 @@ class ConvInc(Operator):
         Y = signals[self.Y]
         pad = self.conv.padding.upper()
         stride = self.conv.strides
+        output_spatial_shape = (
+            self.conv.output_shape.spatial_shape if self.transpose else None
+        )
 
         X = X.reshape(self.conv.input_shape.shape)
         Y = Y.reshape(self.conv.output_shape.shape)
@@ -199,14 +219,115 @@ class ConvInc(Operator):
             W = W[None, :, :, :]
             Y = Y[None, :, :]
             stride = (1,) + stride
+            if output_spatial_shape is not None:
+                output_spatial_shape = (1,) + output_spatial_shape
 
         # add empty batch dimension
         X = X[None, ...]
 
-        def step_conv():
-            Y[...] += conv2d.conv2d(X, W, pad=pad, stride=stride)[0]
+        if self.transpose:
 
-        return step_conv
+            def step_conv_transpose():
+                Y[...] += conv2d.conv2d_gradx(
+                    W, X, xsize=output_spatial_shape, pad=pad, stride=stride
+                )[0]
+
+            return step_conv_transpose
+
+        else:
+
+            def step_conv():
+                Y[...] += conv2d.conv2d(X, W, pad=pad, stride=stride)[0]
+
+            return step_conv
+
+
+class ConvInc(GeneralConvInc):
+    """Apply convolutional weights to input signal.
+
+    .. versionadded:: 3.0.0
+
+    Parameters
+    ----------
+    W : Signal
+        The convolutional weights (a.k.a. the kernel).
+    X : Signal
+        The input signal.
+    Y : Signal
+        Output signal to be incremented.
+    conv : `~nengo.Convolution`
+        The Convolution transform being applied.
+    tag : str, optional
+        A label associated with the operator, for debugging purposes.
+
+    Attributes
+    ----------
+    W : Signal
+        The convolutional weights.
+    X : Signal
+        The input signal.
+    Y : Signal
+        Output signal to be incremented.
+    conv : `~nengo.Convolution`
+        The Convolution transform being applied.
+    tag : str, optional
+        A label associated with the operator, for debugging purposes.
+
+    Notes
+    -----
+    1. sets ``[]``
+    2. incs ``[Y]``
+    3. reads ``[W, X]``
+    4. updates ``[]``
+    """
+
+    def __init__(self, W, X, Y, conv, tag=None):
+        super().__init__(W, X, Y, conv, tag=tag)
+        assert not self.transpose
+
+
+class ConvTransposeInc(GeneralConvInc):
+    """Apply transposed convolutional weights to input signal.
+
+    .. versionadded:: 3.2.0
+
+    Parameters
+    ----------
+    W : Signal
+        The convolutional weights (a.k.a. the kernel).
+    X : Signal
+        The input signal.
+    Y : Signal
+        Output signal to be incremented.
+    conv : `~nengo.transforms.ConvolutionTranspose`
+        The ConvolutionTranspose transform being applied.
+    tag : str, optional
+        A label associated with the operator, for debugging purposes.
+
+    Attributes
+    ----------
+    W : Signal
+        The convolutional weights.
+    X : Signal
+        The input signal.
+    Y : Signal
+        Output signal to be incremented.
+    conv : `~nengo.transforms.ConvolutionTranspose`
+        The ConvolutionTranspose transform being applied.
+    tag : str, optional
+        A label associated with the operator, for debugging purposes.
+
+    Notes
+    -----
+    1. sets ``[]``
+    2. incs ``[Y]``
+    3. reads ``[W, X]``
+    4. updates ``[]``
+    """
+
+    def __init__(self, W, X, Y, conv, tag=None):
+        super().__init__(W, X, Y, conv, tag=tag)
+        assert self.transpose
 
 
 @Builder.register(NoTransform)